components/MaterialSimulationWizard.py: minor
git-svn-id: svn://danse.us/VNET/vnf/trunk@216 056da4b6-1853-4d91-9082-68a8d0adb2e4

+++ vnf/components/MaterialSimulationWizard.py
@@ -238,10 +238,9 @@ class MaterialSimulationWizard(base):
         except AuthenticationError, err:
             return err.page 
 
-        # remove this experiment
-        simulation = director.clerk.getRecordByID(
-            self.inventory.type, self.inventory.id)
-        director.clerk.deleteRecord(simulation)
+        simulation = self._getSimulation(director)
+        # remove this simulation
+        if simulation: director.clerk.deleteRecord(simulation)
         
         # redirect
         actor = 'materialsimulation'; routine = 'listall'
